+++ src/seed_generation/spoilerlog.rs
@@ -2,12 +2,14 @@ use std::{io, path::PathBuf};
 
 use serde_json::{json, to_string_pretty};
 
-use crate::seed_generation::{game_world::{BattleArenaRewards, BossGarage, BossRaceRewards, GemCupRewards, RelicRaceOnlyRewards, Rewards, TokensAndRelicRewards, TrophyRaceRewards, WarpPad}, randomization_datastructures::{GameSetup, UnlockRequirement, UnlockRequirementItem}};
+use crate::seed_generation::seed_settings::{RewardShuffle, WarppadShuffle};
+use crate::seed_generation::{game_world::{BattleArenaRewards, BossGarage, BossRaceRewards, GemCupRewards, RelicRaceOnlyRewards, Rewards, TokensAndRelicRewards, TrophyRaceRewards, WarpPad}, randomization_datastructures::{GameSetup, UnlockRequirement, UnlockRequirementItem}, seed_settings::SeedSettings};
 
 pub fn write_spoilerlog(
     new_rom_path: PathBuf,
     game_setup: GameSetup,
     seed: u32,
+    chosen_settings: &SeedSettings,
 ) -> Result<(), io::Error> {
     let game_world = game_setup.game_world;
     let spoilerlog = json!({
@@ -56,6 +58,30 @@ pub fn write_spoilerlog(
                 "boss_garage": get_formatted_bossgarage(game_world.gemstone_valley.boss_garage)
             },
         },
+        "settings": {
+            "randomization": {
+                "shuffle_adventure": chosen_settings.randomization.shuffle_adventure.to_string(),
+                "shuffle_race_rewards": get_serialized_optional_setting(OptionalSettings::RewardShuffle(chosen_settings.randomization.shuffle_race_rewards)),
+                "warppad_shuffle": get_serialized_optional_setting(OptionalSettings::WarppadShuffle(chosen_settings.randomization.warppad_shuffle)),
+                "warppad_unlock_requirements": chosen_settings.randomization.warppad_unlock_requirements.to_string(),
+                "bossgarage_unlock_requirements": chosen_settings.randomization.bossgarage_unlock_requirements.to_string(),
+                "autounlock_ctrchallenge_relicrace": chosen_settings.randomization.autounlock_ctrchallenge_relicrace.to_string(),
+            },
+            "general": {
+                "relicrace_required_minimum_time": chosen_settings.general.rr_required_minimum_time.to_string(),
+                "relicrace_require_perfects": chosen_settings.general.rr_require_perfects.to_string(),
+                "oxide_final_challenge_unlock": chosen_settings.general.oxide_final_challenge_unlock.to_string(),
+            },
+            "qol": {
+                "skip_mask_hints": chosen_settings.qol.skip_mask_hints.to_string(),
+                "skip_mask_congrats": chosen_settings.qol.skip_mask_congrats.to_string(),
+                "autoskip_podium_cutscenes": chosen_settings.qol.autoskip_podium_cutscenes.to_string(),
+            },
+            "tricks": {
+                "helper_tiziano": chosen_settings.tricks.helper_tiziano.to_string(),
+                "helper_ta": chosen_settings.tricks.helper_ta.to_string(),
+            }
+        }
     });
 
     let mut spoilerlog_path = new_rom_path;
@@ -185,4 +211,35 @@ fn get_formatted_reward(reward: Rewards) -> serde_json::Value {
             })
         }
     }
+}
+
+fn get_serialized_optional_setting(opt_setting: OptionalSettings) -> serde_json::Value {
+    match opt_setting {
+        OptionalSettings::RewardShuffle(x) => {
+            if let Some(y) = x {
+                let mut json_map: serde_json::Map<String, serde_json::Value> = serde_json::Map::new();
+                json_map.insert("include_keys".to_owned(), serde_json::Value::Bool(y.include_keys));
+                json_map.insert("include_gems".to_owned(), serde_json::Value::Bool(y.include_gems));
+                json_map.insert("include_platinum_relics".to_owned(), serde_json::Value::Bool(y.include_platinum_relics));
+                serde_json::Value::Object(json_map)
+            } else {
+                serde_json::Value::String("false".to_owned())
+            }
+        },
+        OptionalSettings::WarppadShuffle(x) => {
+            if let Some(y) = x {
+                let mut json_map: serde_json::Map<String, serde_json::Value> = serde_json::Map::new();
+                json_map.insert("include_battle_arenas".to_owned(), serde_json::Value::Bool(y.include_battle_arenas));
+                json_map.insert("include_gem_cups".to_owned(), serde_json::Value::Bool(y.include_gem_cups));
+                serde_json::Value::Object(json_map)
+            } else {
+                serde_json::Value::String("false".to_owned())
+            }
+        }
+    }
+}
+
+enum OptionalSettings {
+    RewardShuffle(Option<RewardShuffle>),
+    WarppadShuffle(Option<WarppadShuffle>),
 }